api/requirements.txt
django>=5.0.6,<6.0.0
djangorestframework>=3.15.1,<4.0.0
itsdangerous>=2.2.0,<2.3.0 # Itsdangerous : https://itsdangerous.palletsprojects.com/en/2.2.x/ (Verify email)
gunicorn>=22.0.0,<23.0.0
psycopg2-binary>=2.9.9,<3.0.0
types-psycopg2>=2.9.21.20240417,<3.0.0 # Typing stubs for psycopg2
django-cors-headers==4.4.0
django-extensions>=3.2.3,<3.3.0
djangorestframework-simplejwt>=5.3.1
Werkzeug>=3.0.3,<3.1.0
uvicorn>=0.30.1,<0.31.0
requests
google-generativeai>=0.7.2,<0.8.0 # Gemini AI API
requests_mock>=1.12.1,<1.13.0 # Mocking HTTP requests
boto3>=1.34.136,<1.35.0 # For User profile S3 Upload
pillow>=10.4.0,<10.5.0 # Python Image Library
pytz>=2024.1,<2025.0 # Python timezone database